requirements.txt
torch
torchvision
numpy==2.3.0
Pillow
open-clip-torch
fastapi
uvicorn
python-multipart
streamlit
